% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchbot.R
\name{crunchbot}
\alias{crunchbot}
\title{Run code and cat the result as JSON}
\usage{
crunchbot(expr, ...)
}
\arguments{
\item{expr}{Code to run}

\item{...}{Additional arguments passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}}
}
\value{
Invisibly, the return of \code{expr}
}
\description{
Run code and cat the result as JSON
}
