% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{extractDatasetID}
\alias{extractDatasetID}
\title{Get the dataset ID from a URL}
\usage{
extractDatasetID(url)
}
\arguments{
\item{url}{character vector of request URLs or URL segments}
}
\value{
A character vector of equal length containing the dataset IDs, or
\code{NA} if the url contains no dataset ID.
}
\description{
Get the dataset ID from a URL
}
\examples{
extractDatasetID("http://crunch.io/datasets/8159d0c4e26fef8ea371a2d9338ceb91/")
is.na(extractDatasetID("http://crunch.io/users/"))
}
