% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{featureFlags}
\alias{featureFlags}
\alias{featureFlags<-}
\title{View and set feature flags for users}
\usage{
featureFlags(x)

featureFlags(x) <- value
}
\arguments{
\item{x}{a user record (list)}

\item{value}{a list of new attributes to update with. Magically works with
\code{$$}. See examples.}
}
\value{
Getter returns the flag elements from the user preferences. Setter
returns the user record.
}
\description{
View and set feature flags for users
}
\examples{
\dontrun{
featureFlags(u)$notebooks <- TRUE
}
}
