% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{getDatasets}
\alias{getDatasets}
\title{Query the datasets}
\usage{
getDatasets(...)
}
\arguments{
\item{...}{Optional query parameters. Valid fields include "name", "dsid"
(the ID of a dataset), "email" of the owner.}
}
\value{
A data.frame with 4 columns: name, email, account_id, and id.
}
\description{
Query the datasets
}
