% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow-requests.R
\name{getSlowRequests}
\alias{getSlowRequests}
\title{Query the slow requests catalog}
\usage{
getSlowRequests(method = "ANY", path = NULL, threshold = NULL,
  limit = 10)
}
\arguments{
\item{method}{HTTP verb, or "ANY" for any}

\item{path}{URL segment string to filter the results on, e.g. "stream" or "13deb231a"}

\item{threshold}{numeric cutoff value for the timing to include. Only requests taking longer than \code{threshold} will be included}

\item{limit}{integer number of requests to return}
}
\value{
A list of \code{SlowRequest} data.frame objects.
}
\description{
Query the slow requests catalog
}
