% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{getUser}
\alias{getUser}
\title{Get a user record}
\usage{
getUser(id, query)
}
\arguments{
\item{id}{ID of the user to fetch}

\item{query}{Optional search string. The superadmin GUI advises: 'Search
names and emails, and/or make specific queries of Mongo, like
"preferences.labs:true"'. If \code{id} is omitted and a query is provided,
will call \code{getUsers} to search, and will proceed to GET the user record
if the query returns only a single match.}
}
\value{
A list.
}
\description{
Get a user record
}
