% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{getUsers}
\alias{getUsers}
\title{Query the users}
\usage{
getUsers(query = NULL)
}
\arguments{
\item{query}{Optional search string. The superadmin GUI advises: 'Search
names and emails, and/or make specific queries of Mongo, like
"preferences.labs:true"'}
}
\value{
A data.frame with 4 columns: name, email, account_id, and id.
}
\description{
Query the users
}

