% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\alias{unpin}
\title{Pin or unpin a dataset}
\usage{
pin(dataset_id)

unpin(dataset_id)
}
\arguments{
\item{dataset_id}{URL or id of a dataset}
}
\value{
invisibly, the id of the dataset pinned or unpinned
}
\description{
Pin or unpin a dataset
}

