% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow-requests.R
\name{SlowRequest}
\alias{SlowRequest}
\title{Construct a SlowRequest data.frame}
\usage{
SlowRequest(x)
}
\arguments{
\item{x}{a log entry as returned by \code{getSlowRequests()}}
}
\value{
A \code{SlowRequest} object
}
\description{
Construct a SlowRequest data.frame
}
\keyword{internal}
