% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{superadminURL}
\alias{superadminURL}
\title{Construct a superadmin URL}
\usage{
superadminURL(path, host = paste0("http://localhost:",
  getOption("superadmin.local.port")))
}
\arguments{
\item{path}{character relative path}

\item{host}{character path that \code{path} is relative to}
}
\value{
A valid URL.
}
\description{
Construct a superadmin URL
}
