% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnels.R
\name{superConnect}
\alias{superConnect}
\alias{superDisconnect}
\title{Connect and Disconnect to the Admin Server}
\usage{
superConnect(host = "eu",
  local.port = getOption("superadmin.local.port"), remote.port = 8081)

superDisconnect(local.port = getOption("superadmin.local.port"))
}
\arguments{
\item{host}{character, default is "eu" for production. Alternative is
"alpha"}

\item{local.port}{What port to use on your local host}

\item{remote.port}{What port the admin server is on}
}
\value{
Whatever exit status that \code{system2} returns.
}
\description{
\code{superConnect} sets up an ssh tunnel. \code{superDisconnect} closes the tunnel.
You shouldn't need to call either directly: \code{superConnect} will be called
for you when you attempt to get a superadmin URL if the tunnel is not already
set up, and \code{superDisconnect} is called automatically on process exit.
}
