% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{superGET}
\alias{superGET}
\alias{superPOST}
\title{HTTP for superadmin}
\usage{
superGET(...)

superPOST(...)
}
\arguments{
\item{...}{Arguments passed to httr VERBs}
}
\value{
Whatever those HTTP requests return.
}
\description{
These functions auto-tunnel to the superadmin host:port if you're not already
connected.
}
